refactor dylib_applet
